Fixup - Update server_to_sql
